requirements.txt
# For scraping RSS feeds and cleaning HTML
feedparser
beautifulsoup4

# For building the Knowledge Graph
rdflib
# For the LLM pipeline
langchain
langchain-google-genai
langchain-openai
langchain-community
pydantic

#dotenv
python-dotenv

# For the Query API
fastapi
uvicorn[standard]

# For the Event-Driven Pipeline
SPARQLWrapper
redis
